#1 Pass selected model to ollama dynamically
- Now each implementation of ImageProcessor tests whether it can process a request. The first eligible implementation is used.
